Add mulicolor support.
In reference to GH-19 and general discussion, this is the first attempt
on auto-generated color versions.

To make this happen, src/base.ink.svg was modified.
The first example colorset was added as gradiens with one stop.
For easy manipulation, the style attribut was altered.
Please always use a seperate 'FILL' attribute within the star paths.

A second variable in the Makefile ('COLOR') allows for user selection.

'changeStarColor.py' is a script, which reads the file.
It than searches for a hard-coded id and replaces the 'FILL' content.